+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * X axis: time.
   * Y axis: artist playcount.
   * Color: album playcount.
+* Summary: period, total numbers.
 * Highlighting on selection:
   * An info box with date, artist, album and track.
   * Different color for all scrobbles of a given artist.